+++ application/forms/__init__.py
@@ -1,3 +1,18 @@
 from .form import *
 from .childminder import *
 from .account_selection import *
+from .childcare_type import *
+from .contact_info import *
+from .dbs import *
+from .declaration import *
+from .eyfs import *
+from .first_aid import *
+from .hbd import *
+from .magic_link import *
+from .other_people import *
+from .payment import *
+from .personal_details import *
+from .personal_details_address import *
+from .personal_details_childcare import *
+from .references import *
+from .security_question import *
